+++ src/main/java/Game/ClassCharacters/Guerreiro.java
@@ -4,16 +4,16 @@ import java.util.Arrays;
 
 public class Guerreiro extends Personagens {
     String Classe;
-    private double porcentagemXpGanho;
-    private double velocidadeAtaque;
-    private double tempoConjuracaoHabilidade;
-    private double alcanceAtaque;
-    private double tempoRecargaHabilidades;
+    private int porcentagemXpGanho;
+    private int velocidadeAtaque;
+    private int tempoConjuracaoHabilidade;
+    private int alcanceAtaque;
+    private int tempoRecargaHabilidades;
     private String[] habilidades;
 
-    public Guerreiro(String nome, Double vida, Double mana, Double ataque, Double defesa,Double regenVida, Double regenMana,Double velocidade, double porcentagemXpGanho,
-                     double velocidadeAtaque, double tempoConjuracaoHabilidade, 
-                     double alcanceAtaque, double tempoRecargaHabilidades, String[] habilidades) {
+    public Guerreiro(String nome, int vida, int mana, int ataque, int defesa,int regenVida, int regenMana,int velocidade, int porcentagemXpGanho,
+                     int velocidadeAtaque, int tempoConjuracaoHabilidade, 
+                     int alcanceAtaque, int tempoRecargaHabilidades, String[] habilidades) {
         super(nome, vida, mana, ataque, defesa, regenVida,regenMana, velocidade);
         this.Classe = "Guerreiro";
         this.porcentagemXpGanho = porcentagemXpGanho;
@@ -37,23 +37,23 @@ public class Guerreiro extends Personagens {
 
         }
 
-    public double getPorcentagemXpGanho() {
+    public int getPorcentagemXpGanho() {
         return porcentagemXpGanho;
     }
 
-    public double getVelocidadeAtaque() {
+    public int getVelocidadeAtaque() {
         return velocidadeAtaque;
     }
 
-    public double getTempoConjuracaoHabilidade() {
+    public int getTempoConjuracaoHabilidade() {
         return tempoConjuracaoHabilidade;
     }
 
-    public double getAlcanceAtaque() {
+    public int getAlcanceAtaque() {
         return alcanceAtaque;
     }
 
-    public double getTempoRecargaHabilidades() {
+    public int getTempoRecargaHabilidades() {
         return tempoRecargaHabilidades;
     }
