+++ src/main/java/Game/ClassCharacters/Monge.java
@@ -4,17 +4,17 @@ import java.util.Arrays;
 
 public class Monge extends Personagens {
     String Classe;
-    private double porcentagemXpGanho;
-    private double velocidadeAtaque;
-    private double tempoConjuracaoHabilidade;
-    private double alcanceAtaque;
-    private double tempoRecargaHabilidades;
+    private int porcentagemXpGanho;
+    private int velocidadeAtaque;
+    private int tempoConjuracaoHabilidade;
+    private int alcanceAtaque;
+    private int tempoRecargaHabilidades;
     private String[] habilidades;
-    private double fe; // Atributo extra de fé
+    private int fe; // Atributo extra de fé
 
-    public Monge(String nome, Double vida, Double mana, Double ataque, Double defesa,Double regenVida, Double regenMana,Double velocidade, double porcentagemXpGanho,
-                 double velocidadeAtaque, double tempoConjuracaoHabilidade, 
-                 double alcanceAtaque, double tempoRecargaHabilidades, String[] habilidades, double fe) {
+    public Monge(String nome, int vida, int mana, int ataque, int defesa,int regenVida, int regenMana,int velocidade, int porcentagemXpGanho,
+                 int velocidadeAtaque, int tempoConjuracaoHabilidade, 
+                 int alcanceAtaque, int tempoRecargaHabilidades, String[] habilidades, int fe) {
         super(nome, vida, mana, ataque, defesa, regenVida,regenMana,velocidade);
         this.Classe = "Monge";
         this.porcentagemXpGanho = porcentagemXpGanho;
@@ -41,23 +41,23 @@ public class Monge extends Personagens {
         }
 
     // Getters e setters
-    public double getPorcentagemXpGanho() {
+    public int getPorcentagemXpGanho() {
         return porcentagemXpGanho;
     }
 
-    public double getVelocidadeAtaque() {
+    public int getVelocidadeAtaque() {
         return velocidadeAtaque;
     }
 
-    public double getTempoConjuracaoHabilidade() {
+    public int getTempoConjuracaoHabilidade() {
         return tempoConjuracaoHabilidade;
     }
 
-    public double getAlcanceAtaque() {
+    public int getAlcanceAtaque() {
         return alcanceAtaque;
     }
 
-    public double getTempoRecargaHabilidades() {
+    public int getTempoRecargaHabilidades() {
         return tempoRecargaHabilidades;
     }
 
@@ -65,7 +65,7 @@ public class Monge extends Personagens {
         return habilidades;
     }
 
-    public double getFe() {
+    public int getFe() {
         return fe;
     }
 }
